updating img with word
