+++ worddeposit/Controllers/WordCollectionViewCell.swift
@@ -99,18 +99,9 @@ class WordCollectionViewCell: UICollectionViewCell {
     }
     
     @IBAction func wordImageButtonTouched(_ sender: UIButton) {
-        print(word.imgUrl)
-        var config = YPImagePickerConfiguration()
-        config.onlySquareImagesFromCamera = true
-        config.shouldSaveNewPicturesToAlbum = true
-        config.screens = [.library, .photo]
-        config.albumName = "WordDeposit"
-        config.showsPhotoFilters = false
-        
-        let newCapturePhotoImage = UIImage(systemName: "largecircle.fill.circle")?.withTintColor(UIColor.label) ?? config.icons.capturePhotoImage
-        config.icons.capturePhotoImage = newCapturePhotoImage
-        
-        let picker = YPImagePicker(configuration: config)
+        let ypConfig = YPImagePickerConfig()
+        let picker = YPImagePicker(configuration: ypConfig.defaultConfig())
+
         picker.didFinishPicking { (items, true) in
             if let photo = items.singlePhoto {
                 self.isImageSet = true
@@ -151,7 +142,7 @@ class WordCollectionViewCell: UICollectionViewCell {
         
         if isImageSet {
             // if only image has been changed?
-            uploadImage(userId: user.uid, preparedWord: word)
+            uploadImage(userId: user.uid, updatedWord: word)
         } else {
             uploadWord(updatedWord)
         }
@@ -159,7 +150,7 @@ class WordCollectionViewCell: UICollectionViewCell {
     
     /* ********* */
     // check
-    func uploadImage(userId: String, preparedWord: Word) {
+    func uploadImage(userId: String, updatedWord: Word) {
         
         guard let image = wordImageButton.imageView?.image else {
             self.delegate?.showAlert(title: "Error", message: "Fields cannot be empty")
@@ -178,12 +169,12 @@ class WordCollectionViewCell: UICollectionViewCell {
             }
         }
         
-        var preparedWord = preparedWord // convert let to var
+        var updatedWord = updatedWord // convert let to var
         
         let resizedImg = image.resized(toWidth: 400.0)
         guard let imageData = resizedImg?.jpegData(compressionQuality: 0.5) else { return }
         
-        let imageRef = Storage.storage().reference().child("/\(userId)/\(word.id).jpg")
+        let imageRef = Storage.storage().reference().child("/\(userId)/\(updatedWord.id).jpg")
         let metadata = StorageMetadata()
         metadata.contentType = "image/jpg"
         
@@ -201,12 +192,11 @@ class WordCollectionViewCell: UICollectionViewCell {
                     return
                 }
                 guard let url = url else { return }
-                preparedWord.imgUrl = url.absoluteString
+                updatedWord.imgUrl = url.absoluteString
                 
-                if preparedWord.example != self.word.example || preparedWord.translation != self.word.translation {
-                    print("upload word from uploading image")
-                    self.uploadWord(preparedWord)
-                }
+                // updating wordUrl
+                // TODO - should have to own func to update just url may be
+                self.uploadWord(updatedWord)
             }
         }
     }
@@ -230,6 +220,5 @@ class WordCollectionViewCell: UICollectionViewCell {
         self.setupWord(word)
         hideButtons(true)
         isImageSet = false
-//        self.delegate?.showAlert(title: "Cancel Pressed", message: "Word has been updated")
     }
 }
